+++ packages/xchain-ethereum/src/client.ts
@@ -94,7 +94,7 @@ export type EthereumClientParams = XChainClientParams & {
  */
 export default class Client extends BaseXChainClient implements XChainClient, EthereumClient {
   private ethNetwork: EthNetwork
-  private hdNode!: HDNode
+  private hdNode?: HDNode
   private etherscanApiKey?: string
   private explorerUrl: ExplorerUrl
   private infuraCreds: InfuraCreds | undefined
@@ -122,14 +122,13 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
   }: EthereumClientParams) {
     super(Chain.Ethereum, { network, rootDerivationPaths })
     this.ethNetwork = xchainNetworkToEths(network)
-    this.rootDerivationPaths = rootDerivationPaths
     this.infuraCreds = infuraCreds
     this.etherscanApiKey = etherscanApiKey
     this.ethplorerUrl = ethplorerUrl
     this.ethplorerApiKey = ethplorerApiKey
     this.explorerUrl = explorerUrl || this.getDefaultExplorerURL()
     this.setupProviders()
-    this.setPhrase(phrase)
+    phrase && this.setPhrase(phrase)
   }
 
   /**
@@ -139,7 +138,7 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
    */
   purgeClient(): void {
     super.purgeClient()
-    this.hdNode = HDNode.fromMnemonic('')
+    this.hdNode = undefined
   }
 
   /**
@@ -165,6 +164,9 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
     if (walletIndex < 0) {
       throw new Error('index must be greater than zero')
     }
+    if (!this.hdNode) {
+      throw new Error('Phrase must be provided')
+    }
     return this.hdNode.derivePath(this.getFullDerivationPath(walletIndex)).address.toLowerCase()
   }
 
@@ -178,6 +180,9 @@ export default class Client extends BaseXChainClient implements XChainClient, Et
    * Thrown if phrase has not been set before. A phrase is needed to create a wallet and to derive an address from it.
    */
   getWallet(walletIndex = 0): ethers.Wallet {
+    if (!this.hdNode) {
+      throw new Error('Phrase must be provided')
+    }
     return new Wallet(this.hdNode.derivePath(this.getFullDerivationPath(walletIndex))).connect(this.getProvider())
   }
   setupProviders(): void {